Last minute fixes (#1)
* Fix a "Grade" -> "Gradle" typo

@kklamberty caught this typo. Duh.

* Update repo name

The "master" version of the repository is called `intro-to-git`, but the version we'll be using in class is `lab-0-intro-to-git-everyone`, which required one change.

* Fix `gradle` -> `gradlew` typo

Another good catch by @kklamberty.

* Update VS Code commands

Some of the menu items were a little different than as described in the write-up. This tries to bring those closer together.

* Remove the Gradle Tasks extension

This appears to break in the lab for reasons we don't really understand (but which @floogulinc warned me about), so I'm just removing all that.